Python/C split
It's another one of those major changes Brian is famous for! Here, I've split
the Python C code from the ordinary C code. This helps in naming the files and
trying to keep each side of the project organized, but it also lets me experiment
with the scrict-aliasing switch in gcc, which I'm forced to turn off for Python
code.