Move format properties to the StreamFormat class where they belong.
Also change source->stream in Canvas UI as an experiment in better
naming. "Source" is just awfully generic; "stream" and "container"
might be better names regardless of whether it's really stream-like
(all the "streams" in this library are really random-access) or
container-like (in the FFmpeg file sense; as a holder of streams,
it might work well for takes).