Make workspace more general to support audio.
I had been thinking of building a separate audio_workspace, but it turns out
that's not necessary. All the "workspace" does is build a composite_list.
What's really needed is a way to do that for audio, which just requires that
the original current_frame be a range, start_frame to end_frame. With that
change, audio composite_lists can be built easily, and video just becomes
the special case start_frame == end_frame.