workspace: Reduce mutex responsibilities
Be more explicit about what mutex does; it's meant to allow multiple threads
to "read" from the workspace simultaneously, such that, if the caller wants,
the caller can protect its operations with a read/write lock. With this
understanding, the time we spend locked in get_frame_f32 can be reduced
drastically.