FIX: widget_gl drawOneFrame race condition
We were resetting the boolean drawOneFrame, but then calling the
playSingleFrame callback again, causing us to move to the next buffer and
scramble everything. This version implements drawOneFrame as more of a
reference count.

There are probably more race conditions and lock contentions hiding in
this code. Working on a slow system, a lot of things slow to a crawl when
the scene graph is updated. Someday I'll come back to restructure widget_gl.