Canvas UI: Separate rendering from movement
Finally got my head around how to separate these two concepts so that I
could inherit the drawing capability without all of the clip-specific stuff.

The solution moves the common clip-like functionality to a class SceneItem,
where painting takes place. Painting is done with an instance of a painter,
which it just so happens that ThumbnailPainter is.

Clip items are dealt with specifically in ClipItem, and it can now switch
between video and audio, so VideoClip, VideoItem, AudioClip, and AudioItem
all go away.