Sequence: Support negative transition_lengths for gaps
It occurs to me that it's silly to say that there should be no gaps in
sequences. That would make things a little difficult to manage sequences
when you're trying to create a gap into which to insert something. So
here's a simple patch to enable (without support in the editor) gaps using
negative transition_lengths.