Take into account more recent FFmpeg/libav changes
This is a pain to do, mainly because of the warning=error requirement I've
made in the build, but it does help us get version usage something close to
exactly right.

For people doing this in the future: the version numbers you see come from
libav's APIchanges file, which shows when functions got deprecated.