Plugins WIP
Work-in-progress with a fairly spiffy interface for activating and
deactivating plugins. A lot of work to help me learn the Qt4 style,
painter, and font systems.

This commit also makes the first use of QSettings for storing user
settings. I think it's safe to say I won't store all things there. I expect
I'll want some of the plugin settings to be local to the project. But I
don't yet know how I want to do that.