Add a rudimentary alert system and change several methods to properties
I know, I know, mixed update. Mainly this adds an alert system (currently
"Notifications" in the UI; it was that in the code, too, but I got tired of
typing it). This also changes some of the Plugin basics to attributes or
properties, so that they can be implemented as class-level attributes or
whatever.