Big changes: Alerts and sources via plugins
It's time to admit that I lost track of what change preceded what a long
time ago. This is a big commit that *almost* has the editor fully working
again, just missing thumbnails as far as I know. That has to do with an
incomplete change in format classes.

This commit has sources and streams produce alerts that flow through the
composition to tell the user why something failed, and hopefully give them
something to do about it (not yet implemented). It also has sources pulled
up from plugins, and found through connectors, which are convenience
classes that simplify the error handling for resolving a stream reference
into a stream.

More to come.