Rewrite Libav-DV into separate source and codec plugins
This commit is awfully haphazard, since I did it in a hurry and on a train.
A few things I'll need to codify: How to find the name of a stream ("name"
attribute on stream I think will do), the codec of a stream ("codec" on
stream?), and how to report errors in finding the codec. I think a blank
stream with an error message might do the trick, but it'll have to be a
passthrough of some kind. Probably the same sort of connector like with the
VideoStreamRefConnector.