Modify the build just enough to build the Qt extension with PySide
The Debian-based distro I'm working on right now doesn't have the packages
for PyQt4 OpenGL in Python 3, but it does have the packages for PySide. I'm
going to give that a try.

Please don't complain about the hard-coded paths in the build. The
pkg-config files for pyside reference Python 2.7, and even though Python 3
libraries are available, I don't see a way to get their location from the
system. I just have to wait until Debian notices that.