Unslot the _AlertTracker and fix a bug of missing attributes
Truth be told, I don't know why the attributes are missing. All I know is
that when stop_tracking is called somewhere during program shutdown, the
attributes aren't there. So I figure that if I can't find them, the garbage
collector probably got to them first-- which to me is a violation of
expectations, but it's teardown code, so I'll deal.