Start to define model changes in terms of QUndoCommands
I don't know why this didn't occur to me before. This will be much simpler
and much more manageable. The idea is to decompose the whole ItemManipulator
into much, much smaller commands, such as this first one,
_DeleteItemsFromSequence. Then each command is responsible for the state
necessary to restore things before that command.

Where it should get interesting, and much easier to understand, is when a
set of items is being moved around in a sequence. Right now, I make a map
of all the valid places to set items in a sequence, then keep that map
around and do some complicated math on it. With this model, we should be
able to check, piecewise, if the next move is a valid move, and with any
luck, be able to combine it with the last move on the stack.