Implement the ClipManipulator with undo commands
Really, this is an awful way to do it, but I implement the old guarantees
about state to get the tests to pass. And that's just an exercise, because
really I think this is the wrong way to go about it.

I think what I'll do is remove the can_set_* methods and change the
expectations on the set_* methods such that if setting fails, your clip's
status is indeterminate (in which case you should probably call set_space)
or it is missing from the space (in which case, you should pretend the item
is still in the user's hand waiting to be put down).