Accept a floating-point X for manipulators so we can do the rounding ourselves
This will come into play when we're moving items with different rates to
them. When it comes time to implement that in the main ItemManipulator, the
plan is to (a) move the primary item, (b) look at its final position, and
then (c) move other items to match its final position. (Otherwise, you'll
move maybe a primary video and audio items around it will try to sneak
around as you're moving the cursor, even if the primary isn't moving.)