Canvas: Thumbnails are now cached
There are a few issues yet to be dealt with: How to deal with the size of a
thumbnail (should they all be the same size and just be scaled later?),
dealing with streams that change (such as the thumbnails of a collapsed
sequence, or heaven help us if we show post-processed thumbnails), two
streams requesting the same uncached frame (only one will get notified when
it's available).

Still, this greatly improves the responsiveness of the program.