Put the clock in a UIManager class
The UIManager class will be home base for anything needing the standard
services provided by the UI (such as the clock and the asset list). For
those extending the interface, there will be some classes they can extend.

Part of the purpose of hiding the clock in the UIManager is to have a
central place to swap it out if the user wants to change clock sources
(which right now is controlled by audio output).