Add another layer of abstraction: assets vs. sources
I've decided that sources should not, themselves, be assets. Assets can be
things other than sources. This should open it up in the future if people
want to store effects templates, scripts, etc. in the asset list.

Spaces also cease to be sources themselves. They didn't even implement the
protocol for sources. The SpaceAsset takes over that job by taking a space
and creating a source for it on request.