I WIN! So much win! Dragging with two-way anchors
This does the dragging with two-way anchors among top-level items. The
scene also places its focus item first so that timing is based on that
item. This means sound items can be placed with much accuracy, while
keeping their video items largely in sync.

Although, now that I think about it, that's probably the right way to
handle one-way handles, but it might be the wrong way to handle two-way.
Someone might want their video and audio to stay perfectly in sync. I
should probably make that an option.