Enable GL mode for the DVReconstructionFilter
This is three out of the three main filters I needed to add GL support to.
This finally gives most of the path exercised by the editor GL support.

I'm developing this on my laptop right now, which doesn't have the hardware
capabilities needed to run this mode natively, so I'm running a copy of
Mesa with the LLVM-based Gallium drivers, which compiles the GL shaders so
that they can run on the CPU. I expected their implementation to be more
optimized than mine, but my software implementation was actually faster,
probably because I took more advantage of lookup tables. Who knows? But
with capable hardware, the GL version should run much, much faster.