Start to convert Codec with model/view separation
This was my revelation about not trying to push parameters into the stream
instances themselves. We could have multiple copies of a codec stream in
different states of following the model. Trying to have the model *in* the
codec stream... well, you'd have to copy any parameter changes between them
yourself, wouldn't you? And that's just dumb.