gl: Add API for holding a ready GL context in reserve for the current thread
I'm not sure what the final use of this API will be. I'm currently using it
to let me do GL-based operations in a pure Python app.

One major downside of it is that it opens a new X connection and makes a
brand new GL context per thread, but that probably just means I need to do
better thread management.