x264: Add a module for doing H.264 encoding with libx264
This just adds the beginnings of a libx264-based encoder to the mix, as
well as use it in the sample program decode_raw_image.py (which should
probably be split with a different name). Like the MPEG2 subsampler, this
assumes a great deal: a 720x480 4:2:0 YUV planar frame with a 16:9 SAR,
studio levels, interlacing, Rec. 709 colors, 29.97 frame rate, a constant
ratefactor rate control locked at 23, and Annex B encoding.

The most likely thing that will happen is pulling out the params structure
as its own Python object, and exposing all of the options as attributes.

Also the resulting stream from all of this works, but not well. I'll have
to experiment more.