libac: AVAudioDecoder: Add timestamp maps
This is a fix to a problem when decoding audio in DV (and any other
container where the timestamps may be inaccurate). When trust_timestamps is
False (the default), the AVAudioDecoder decodes the stream from the
beginning and builds a map of timestamps to samples. The timestamp of the
first packet is taken verbatim, but each packet thereafter is decoded to
see exactly how many samples are in it, and the start and duration stamps
of all packets afterwards are adjusted to the exact sample.

This might not be the final patch. My tests show that there is a little
static noise still left over, so I'm going to see where the difference is.