Add module for writing Matroska files
This gives support for writing very simple Matroska files, as well as a
utility function for writing out PCM float audio files in Matroska. (The
utility function, while a great example for how to use the MatroskaWriter
class, is awfully slow because it processes each sample in Python. A big
improvement could be achieved by accessing the audio frame as a buffer.)