gl: Make utility functions for more modern, efficient use of GL
New routines to separately compile and link shaders, a standard vertex
shader that fixes the surprising issue that most (if not all!) of my
existing fragment shaders didn't take the texture->frame offset into
account, use of vertex buffers and generic attributes instead of glBegin,
glTexCoord, glVertex, and glEnd, holding onto the framebuffer, using
GL_TRIANGLE_FAN instead of GL_QUADS-- all of these things make the GL usage
more compatible with OpenGL 3.2 and above, which I hope to be able to use
in this project someday.

An aside: MESA_DEBUG=1 is an awesome, awesome feature of Mesa that lets me
know when I've got GL errors. It unfortunately doesn't say where, but it's
easy to do that with gdb and setting a breakpoint on _mesa_error.