gl: Add document explaining the default state of the GL
Future (with respect to 2.1) versions of GL do not support saving the GL
state, which is a bit of a relief to me, because it's hard to make sure the
right things are saved anyways. The alternative for us is to set a good set
of defaults, and always revert to them before the end of the function. This
document describes those defaults.