process: Don't fire up a context by default
Instead of doing this at module load, defer it until either the first video
rendering or when the user calls check_context_supported().

Because of the way GLEW works, this isn't foolproof. It's perfectly
possible to have two contexts with completely different capabilities in the
same process, and GLEW will only detect capabilities on one of them.
Eventually I'll switch to GLEW MX, which will fix that issue.