Add the video encoder script I'm using for my private project
This uses the libdv audio decoder right now because I'm testing it. The
libdv decoder is still producing a big time offset after errors, and worse,
it goes silent after fifteen or twenty minutes, so I'm going to have to
diagnose that. The first problem may just be down to re-syncing to the
video PTS on errors.